requirements.txt
openai
langchain==0.1.13
langchain-community==0.0.30
tiktoken
faiss-cpu
PyPDF2
streamlit==1.31.1
python-dotenv
numpy==1.26.4
